Update the migration helper
